+++ src/js/button.js
@@ -55,8 +55,7 @@ class Button extends Component {
   }
 
   buildCSSClass() {
-    // TODO: Change vjs-control to vjs-button?
-    return `vjs-control ${super.buildCSSClass()}`;
+    return `vjs-control vjs-button ${super.buildCSSClass()}`;
   }
 
   // Click - Override with specific functionality for button